+++ NEWS.md
@@ -7,6 +7,8 @@
   iterators (#891).
 - Signatures in attribute names are now parsed correctly (#892).
 - Attribute specifications for `all` are now parsed correctly (#889).
+- Predefined `"="` and `"/="` operators are no longer declared for file
+  types.
 
 ## Version 1.12.2 - 2024-05-15
 - Fixed a crash when `'transaction` is used with a record type.